split agent pages
